Add target to copy README.md during build process
This commit introduces a new target named `CopyReadmeFile` in the `Directory.Build.targets` file. It is configured to run before the `Build` target and copies the `README.md` file from the parent directory to the project's directory. The `SkipUnchangedFiles` attribute is set to `true`, allowing the process to skip copying if the file has not changed.